refactor: Updated project structure
