+++ src/core/rag/stores/database/base_database_store.py
@@ -1,0 +1,5 @@
+from abc import ABC, abstractmethod
+from typing import List
+
+class BaseDatabaseStore(ABC):
+    """Abstract base class for database stores"""
