+++ src/core/rag/stores/vector/base_vector_store.py
@@ -4,14 +4,13 @@ from llama_index.core import VectorStoreIndex, Document
 
 class BaseVectorStore(ABC):
     """Abstract base class for vector stores"""
-    
+
     @abstractmethod
     def create_index(self, documents: List[Document]) -> VectorStoreIndex:
         """Create index from documents"""
         pass
-    
+
     @abstractmethod
     def get_index(self) -> VectorStoreIndex:
         """Get existing index"""
         pass
-
