+++ src/core/rag/stores/vector/pgvector/config.py
